Interpolate values as expressions
esprima is used to parse the interpolated expressions. The required rewriting
is performed to detect which fields to listen to, before getting converted back
into code and written to the compiled result.

`this` in the templates now refers to the scope (the specified data object),
and all interpolated values need to be correctly prefixed (i.e. `this` is
explicit, not implicit).

Net gain is that ornament properly supports nested values, in addition to
understanding single expression statements, e.g. `this.fn(this.value)`.